refactor : 상수 메세지들 변수화
